main_svc: events complete

+++ ewm-service/src/main/java/ru/practicum/ewm/service/DefaultCategoryService.java
@@ -7,15 +7,23 @@ import ru.practicum.ewm.dto.category.CategoryMapper;
 import ru.practicum.ewm.dto.category.NewCategoryDto;
 import ru.practicum.ewm.exception.AlreadyExistException;
 import ru.practicum.ewm.exception.NotFoundException;
+import ru.practicum.ewm.exception.ValidationException;
 import ru.practicum.ewm.model.Category;
 import ru.practicum.ewm.repository.CategoryRepository;
 
+import org.springframework.data.domain.Pageable;
+import ru.practicum.ewm.repository.EventRepository;
+
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class DefaultCategoryService implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
+
     @Override
     public CategoryDto create(NewCategoryDto newCategoryDto) {
         String name = newCategoryDto.getName();
@@ -30,13 +38,12 @@ public class DefaultCategoryService implements CategoryService {
 
     @Override
     public void remove(Long categoryId) {
-
-// TODO Существуют события, связанные с категорией
-
-        categoryRepository.findById(categoryId).
-                orElseThrow(() -> new NotFoundException(Category.class,
-                        String.format(" with id=%d ", categoryId)));
-
+        categoryRepository.findById(categoryId)
+                        .orElseThrow(() -> new NotFoundException(Category.class,
+                                String.format(" with id=%d ", categoryId)));
+        if (!eventRepository.findAllByCategoryId(categoryId).isEmpty()) {
+            throw new ValidationException(Category.class, "Существуют события, связанные с категорией");
+        }
         categoryRepository.deleteById(categoryId);
     }
 
@@ -44,8 +51,8 @@ public class DefaultCategoryService implements CategoryService {
     public CategoryDto update(Long categoryId, NewCategoryDto newCategoryDto) {
         String name = newCategoryDto.getName();
 
-        Category stored = categoryRepository.findById(categoryId).
-                orElseThrow(() -> new NotFoundException(Category.class,
+        Category stored = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new NotFoundException(Category.class,
                         String.format(" with id=%d ", categoryId)));
 
         Optional<Category> existing = categoryRepository.findByName(name);
@@ -57,5 +64,19 @@ public class DefaultCategoryService implements CategoryService {
         stored.setName(name);
 
         return CategoryMapper.toDto(categoryRepository.save(stored));
+    }
+
+    @Override
+    public List<CategoryDto> findAll(Pageable pageable) {
+        return categoryRepository.findAll(pageable).stream()
+                .map(CategoryMapper::toDto)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public CategoryDto findById(Long categoryId) {
+        return CategoryMapper.toDto(categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new NotFoundException(Category.class,
+                        String.format(" with id=%d ", categoryId))));
     }
 }
